Updated to Acid 0.9.8

+++ Acid/include/acid/Inputs/Joysticks.hpp
@@ -41,8 +41,8 @@ namespace acid
 
 		const float *m_axes;
 		const unsigned char *m_buttons;
-		int m_axeCount;
-		int m_buttonCount;
+		uint32_t m_axeCount;
+		uint32_t m_buttonCount;
 	};
 
 	/// <summary>
@@ -58,19 +58,10 @@ namespace acid
 		/// Gets this engine instance.
 		/// </summary>
 		/// <returns> The current module instance. </returns>
-		static Joysticks *Get()
-		{
-			return Engine::Get()->GetModule<Joysticks>();
-		}
+		static Joysticks *Get() { return Engine::Get()->GetModule<Joysticks>(); }
 
-		/// <summary>
-		/// Creates a new joysticks module.
-		/// </summary>
 		Joysticks();
 
-		/// <summary>
-		/// Deconstructor for the joysticks module.
-		/// </summary>
 		~Joysticks();
 
 		void Update() override;
@@ -110,13 +101,13 @@ namespace acid
 		/// </summary>
 		/// <param name="port"> The joystick to the the button count from. </param>
 		/// <returns> The number of buttons the joystick offers. </returns>
-		uint32_t GetCountButtons(const JoystickPort &port) const { return static_cast<uint32_t>(m_connected.at(port).m_buttonCount); }
+		uint32_t GetCountButtons(const JoystickPort &port) const { return m_connected.at(port).m_buttonCount; }
 
 		/// <summary>
 		/// Gets the number of axes the joystick offers.
 		/// </summary>
 		/// <param name="port"> The joystick to the the axis count from. </param>
 		/// <returns> The number of axes the joystick offers. </returns>
-		uint32_t GetCountAxes(const JoystickPort &port) const { return static_cast<uint32_t>(m_connected.at(port).m_axeCount); }
+		uint32_t GetCountAxes(const JoystickPort &port) const { return m_connected.at(port).m_axeCount; }
 	};
 }
